requirements.txt
fastapi==0.135.2
httpx==0.28.1
pydantic==2.12.3
python-dotenv==1.2.2
uvicorn==0.38.0
resend==2.10.0
